requirements.txt
backports.pbkdf2==0.1
capstone==5.0.1
cryptography==41.0.3
dotnetfile==0.2.4
maco==1.1.0
pefile==2019.4.18
pycryptodome==3.20.0
pycryptodomex==3.20.0
python-magic==0.4.27
pythonnet==3.0.3
requests==2.31.0
